+++ src/os_win32.c
@@ -477,8 +477,7 @@ get_exe_name(void)
 
     if (exe_path == NULL && exe_name != NULL)
     {
-	exe_path = vim_strnsave(exe_name,
-				     (int)(gettail_sep(exe_name) - exe_name));
+	exe_path = vim_strnsave(exe_name, gettail_sep(exe_name) - exe_name);
 	if (exe_path != NULL)
 	{
 	    // Append our starting directory to $PATH, so that when doing
